+++ mteb/languages/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from .check_language_code import check_language_code
 from .iso_mappings import ISO_TO_FAM, ISO_TO_FAM_LEVEL0, ISO_TO_LANGUAGE, ISO_TO_SCRIPT
-from .language_script import LanguageScripts
+from .language_scripts import LanguageScripts
 from .programming_languages import PROGRAMMING_LANGS
 
 __all__ = [
