+++ mteb/languages/language_scripts.py
